requirements.txt
apscheduler==3.11.0
atproto==0.0.56
chromadb==0.5.23
fastapi==0.115.6
kafka==1.3.5
kafka_python==2.0.2
python-dotenv==1.0.1
uvicorn==0.32.1
